pyproject.toml
[tool.poetry]
name = "glow-dogs"
version = "0.1.0"
description = ""
authors = ["Sofiia Kashperova <kashperova.study@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "2.4.0"
pytest = "8.3.4"
pre-commit = "4.0.1"
hydra-core = "1.3.2"
tqdm = "4.66.5"
torchvision = "0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
